+++ NEWS.md
@@ -15,6 +15,7 @@ BUG FIXES
 - fix column highlighting in cog table sort/filter when loading from state
 - fix footer histograms when bin start/end values are not integers
 - fix view() not working with name specified only (no state)
+- fix issue with zero values in filter being ignore when restoring state
 
 Version 0.7.9.1
 -------------------------------------------------------------------------------